clean up sources
