--- sources/visualizations.ts
@@ -1,10 +1,0 @@
-import BafinStimmrechte from './bafin-stimmrechte/Component'
-
-const visualizationSources = [
-  {
-    id: 'bafin-stimmrechte',
-    Component: BafinStimmrechte
-  }
-]
-
-export default visualizationSources
